int casting on initial math.ceil call instead of at references to calculated values

+++ zipper.py
@@ -10,10 +10,10 @@ def lzss_bytewise_encode(bits, W, L):
         print("Padding {} bits".format(bits.fill()))
     source_bytes = bytes(bits.tobytes())
     # log2_W is the number of bits needed to represent W
-    log2_W = math.ceil(math.log(W + 1, 2))
+    log2_W = math.ceil(int(math.log(W + 1, 2)))
     WintConverter = "0" + str(int(log2_W)) + "b"
     # log2_L is the number of bits needed to represent L
-    log2_L = math.ceil(math.log(L + 1, 2))
+    log2_L = math.ceil(int(math.log(L + 1, 2)))
     LintConverter = "0" + str(int(log2_L)) + "b"
     # encoding n characters as a tuple requires 1 + log2_W + log2_L + 8 bits
     lempelength = log2_L + log2_W + 9
@@ -49,9 +49,9 @@ def lzss_bytewise_encode(bits, W, L):
 
 def lzss_bytewise_decode(encoded_bits, W, L):
     # log2_W is the number of bits needed to represent W
-    log2_W = math.ceil(math.log(W + 1, 2))
+    log2_W = math.ceil(int(math.log(W + 1, 2)))
     # log2_L is the number of bits needed to represent L
-    log2_L = math.ceil(math.log(L + 1, 2))
+    log2_L = math.ceil(int(math.log(L + 1, 2)))
     # print("log2W = {}, log2L = {}".format(log2_W, log2_L))
     source = bitarray.bitarray()
     readpos = 0
@@ -88,10 +88,10 @@ def lz77_bytewise_encode(bits, W, L):
         print("Padding {} bits".format(bits.fill()))
     source_bytes = bytes(bits.tobytes())
     # log2_W is the number of bits needed to represent W
-    log2_W = math.ceil(math.log(W + 1, 2))
+    log2_W = math.ceil(int(math.log(W + 1, 2)))
     WintConverter = "0" + str(int(log2_W)) + "b"
     # log2_L is the number of bits needed to represent L
-    log2_L = math.ceil(math.log(L + 1, 2))
+    log2_L = math.ceil(int(math.log(L + 1, 2)))
     LintConverter = "0" + str(int(log2_L)) + "b"
     encoded = bitarray.bitarray()
     # pos is bytewise address
@@ -131,10 +131,10 @@ def lz77_n_bitwise_encode(bits, W, L, n):
         print("len(bits) = {} is not divisible by given n = {}. Returned false.".format(len(bits), n))
         return False
     # log2_W is the number of bits needed to represent W
-    log2_W = math.ceil(math.log(W + 1, 2))
+    log2_W = math.ceil(int(math.log(W + 1, 2)))
     WintConverter = "0" + str(int(log2_W)) + "b"
     # log2_L is the number of bits needed to represent L
-    log2_L = math.ceil(math.log(L + 1, 2))
+    log2_L = math.ceil(int(math.log(L + 1, 2)))
     LintConverter = "0" + str(int(log2_L)) + "b"
     encoded = bitarray.bitarray()
     # pos is a bitwise address - translate to indexing address by dividing by n
@@ -176,9 +176,9 @@ def lz77_n_bitwise_encode(bits, W, L, n):
 # decodes for both n_bitwise and bytewise encoders
 def lz77_n_bitwise_decode(encoded_bits, W, L, n):
     # log2_W is the number of bits needed to represent W
-    log2_W = math.ceil(math.log(W + 1, 2))
+    log2_W = math.ceil(int(math.log(W + 1, 2)))
     # log2_L is the number of bits needed to represent L
-    log2_L = math.ceil(math.log(L + 1, 2))
+    log2_L = math.ceil(int(math.log(L + 1, 2)))
     # print("log2W = {}, log2L = {}".format(log2_W, log2_L))
     source = bitarray.bitarray()
     readpos = 0
